Add dropdown-menu component docs
