+++ packages/dropdown-menu/src/components/dropdown-content.md
@@ -1,0 +1,9 @@
+```jsx static
+import { DropdownContent } from "@nteract/dropdown-content"
+```
+
+Display dropdown menu contents
+
+```
+<DropdownContent />
+```
