+++ packages/dropdown-menu/src/components/dropdown-menu.md
@@ -1,0 +1,9 @@
+```jsx static
+import { DropdownMenu } from "@nteract/dropdown-menu"
+```
+
+Display a dropdown menu
+
+```
+<DropdownMenu />
+```
