+++ packages/dropdown-menu/src/components/dropdown-trigger.md
@@ -1,0 +1,9 @@
+```jsx static
+import { DropdownTrigger } from "@nteract/dropdown-trigger"
+```
+
+Display a dropdown trigger
+
+```
+<DropdownTrigger />
+```
